Move functions into utils file
